setup internal otel encode module
