+++ exporter/elasticsearchexporter/internal/otel/otel.go
@@ -1,0 +1,12 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package otel // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/otel"
+
+// TODO: remove const repetition
+const (
+	dataStreamType      = "data_stream.type"
+	dataStreamDataset   = "data_stream.dataset"
+	dataStreamNamespace = "data_stream.namespace"
+	mappingHintsAttrKey = "elasticsearch.mapping.hints"
+)
